+++ src/all/rouvideo/assets/i18n/messages_en.properties
@@ -10,3 +10,8 @@ all_videos=All videos
 latest_recent=Latest
 most_viewed=Most viewed
 most_liked=Most liked
+ChineseAV=Chinese AV
+Madou_Media=Madou Media
+Selfie_leaked=Selfie Leaked
+Tanhua=Escort
+JAV=JAV
