+++ src/all/rouvideo/assets/i18n/messages_vi.properties
@@ -10,3 +10,8 @@ all_videos=Tất cả video
 latest_recent=Mới nhất
 most_viewed=Xem nhiều nhất
 most_liked=Thích nhiều nhất
+ChineseAV=AV Tàu
+Madou_Media=Madou Media
+Selfie_leaked=Clip sex tự quay
+Tanhua=Check hàng
+JAV=JAV
