+++ src/all/rouvideo/assets/i18n/messages_zh.properties
@@ -10,3 +10,8 @@ all_videos=全部視頻
 latest_recent=時間
 most_viewed=觀看
 most_liked=點贊
+ChineseAV=國產AV
+Madou_Media=麻豆傳媒
+Selfie_leaked=自拍流出
+Tanhua=探花
+JAV=日本
