+++ src/all/rouvideo/src/eu/kanade/tachiyomi/animeextension/all/rouvideo/RouVideoFilter.kt
@@ -1,0 +1,58 @@
+package eu.kanade.tachiyomi.animeextension.all.rouvideo
+
+import eu.kanade.tachiyomi.animesource.model.AnimeFilter
+import eu.kanade.tachiyomi.lib.i18n.Intl
+
+internal object RouVideoFilter {
+
+    class CategoryFilter(intl: Intl) : UriPartFilter(
+        intl["category"],
+        arrayOf(
+            Tag(intl["featured"], FEATURED),
+            Tag(intl["other_watching"], WATCHING),
+        )
+            .plus(categories(intl))
+            .plus(Tag(intl["all_videos"], ALL_VIDEOS)),
+    )
+
+    class TagFilter(intl: Intl, tags: Tags) : UriPartFilter(
+        intl["tag"],
+        tags,
+    )
+
+    class SortFilter(intl: Intl) : UriPartFilter(
+        intl["sort"],
+        arrayOf(
+            Pair(intl["latest_recent"], SORT_LATEST_KEY),
+            Pair(intl["most_viewed"], SORT_VIEW_KEY),
+            Pair(intl["most_liked"], SORT_LIKE_KEY),
+        ),
+    )
+
+    open class UriPartFilter(displayName: String, private val options: Tags) :
+        AnimeFilter.Select<String>(displayName, options.map { it.first }.toTypedArray()) {
+        fun toUriPart() = options[state].second
+        fun isEmpty() = options[state].second == ""
+        fun isDefault() = state == 0
+    }
+
+    fun categories(intl: Intl) = setOf(
+        Tag(intl["ChineseAV"], "國產AV"), // ChineseAV
+        Tag(intl["Madou_Media"], "麻豆傳媒"), // Madou Media
+        Tag(intl["Selfie_leaked"], "自拍流出"), // Selfie leaked
+        Tag(intl["Tanhua"], "探花"), // Tanhua (Flower exploration - Thám hoa - Check hàng)
+        Tag("OnlyFans", "OnlyFans"),
+        Tag(intl["JAV"], "日本"), // JAV
+    )
+
+    const val FEATURED = "featured"
+    const val WATCHING = "watching"
+    const val ALL_VIDEOS = "all-videos"
+
+    const val SORT_LATEST_KEY = "createdAt"
+    const val SORT_LIKE_KEY = "likeCount"
+    const val SORT_VIEW_KEY = "viewCount"
+}
+
+typealias Tags = Array<Tag>
+typealias Tag = Pair<String, String>
